첨부파일 업로드 기능 추가(아직 미완성)

+++ src/main/java/com/finalteam5/otisrm/dto/srRequest/SrRqstAtch.java
@@ -9,5 +9,5 @@ public class SrRqstAtch {
 	private String srRqstAtchNm;
 	private String srRqstAtchMimeType;
 	private byte[] srRqstAtchData;
-	private  int srRqstAtchSize;
+	private long srRqstAtchSize;
 }
